requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
six==1.17.0
tzdata==2024.2
Unidecode==1.3.8
Werkzeug==3.1.3
